+++ R/residuals.R
@@ -337,16 +337,6 @@ residuals.sdmTMB <- function(object,
   # retrieve function that called this:
   sys_calls <- unlist(lapply(sys.calls(), deparse))
   visreg_call <- any(grepl("setupV", substr(sys_calls, 1, 7)))
-  if (!visreg_call) {
-    if (type_was_missing || type == "mle-laplace") {
-      msg <- paste0("Note what used to be the default sdmTMB residuals ",
-        "(before version 0.4.3.9005) are now `type = 'mle-eb'`. We recommend using ",
-        "the current default `'mle-mvn'`, which takes one sample from the approximate ",
-        "posterior of the random effects or `dharma_residuals()` using a similar ",
-        "approach.")
-      cli_inform(msg)
-    }
-  }
   if (type == "mle-laplace") type <- "mle-eb"
   model_missing <- FALSE
   if (identical(model, c(1, 2))) model_missing <- TRUE